enhance: Only export one denormalize
BREAKING CHANGE: Remove denormalizeAndTracked
